backend/pyproject.toml
[project]
name = "myrss"
version = "0.1.0"
description = "Self hosted RSS reader"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "opentelemetry-instrumentation-fastapi>=0.58b0",
    "pydantic-settings>=2.10.1",
    "pyyaml>=6.0.2",
    "structlog>=25.4.0",
    "uvicorn>=0.35.0",
    "xmltodict>=0.15.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/myrss"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.13.0",
]
